+++ src/api/services/prediction/index.ts
@@ -1,0 +1,6 @@
+/*
+*   Function
+*   Copyright © 2023 NatML Inc. All Rights Reserved.
+*/
+
+export * from "./service"